// src/types/Rental.ts
export type Rental = {
  check_in_time: string;
  checkout_time: string;
  city: string;
  country: string;
  currency: string;
  id: string;
  image_path: string;
  latitude: number;
  longitude: number;
  name: string;
  object: string;
  postal_code: string;
  rates_confirmed: boolean;
  status: string;
  time_zone: string;
  url: string;
};